Bandaid fix for corrupt entity crashes

+++ src/world.c
@@ -215,6 +215,7 @@ void SV_FindTouchedLeafs(edict_t *ent, mnode_t *node)
 
 void SV_LinkEdict(edict_t *ent, bool touch_triggers)
 {
+	if(((u64)ent->area.prev) & 0xFF00000000000000) return; // corrupt entity
 	if(ent->area.prev) SV_UnlinkEdict(ent); // unlink from old position
 	if(ent == sv.edicts) return; // don't add the world
 	if(ent->free) return;
